[FEATURE] Move TestSystem classes to compat folder
